+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - connection retry using `promise-retry`
 - initialization retry using `promise-retry` to handle concurrect data structure initialization attempts
 - deprecated `settings.connect` in favor of `settings.connectionParams`
+- add `settings.onReady`
 
 ## v2.9.0